+++ CHANGELOG.md
@@ -30,6 +30,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## Added
 
 - Add support for Ruby 3.4.
+- Support response-level `produces` (allows definition of different schemas per content type)
 
 ## Changed